% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{admmlasso_logC}
\alias{admmlasso_logC}
\title{Fit l1-penalized logistic regression model using ADMM}
\usage{
admmlasso_logC(X, y, lam, rho = 0.001, maxit = 1000L, tol = 0.001)
}
\arguments{
\item{X}{Covariate matrix (no column for intercept)}

\item{y}{Vector of observations (coded in -1/1)}

\item{lam}{Tuning parameter for lasso penalty}

\item{rho}{Tuning parameter for ADMM optimization}

\item{maxit}{Maximum number of iterations}

\item{tol}{Convergence criterion}
}
\value{
Vector containing updated estimate of beta vector
}
\description{
Use an ADMM approach to find the parameters for a l1-penalized logistic regression model.
Finds solution to argmin_beta sum(log(1+-yX beta)) + lambda*sum(|beta|)
}
