\name{aeffp-package}
\alias{aeffp-package}
\alias{aeffp}
\docType{package}
\title{
\packageTitle{aeffp}
}
\description{
\packageDescription{aeffp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aeffp}
\packageIndices{aeffp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{aeffp}

Maintainer: \packageMaintainer{aeffp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
