% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_logistic.R
\name{b_update2}
\alias{b_update2}
\title{Update beta estimates using L-BFGS-B}
\usage{
b_update2(X, y, u, z, rho)
}
\arguments{
\item{X}{Covariate matrix (no column for intercept)}

\item{y}{Vector of observations (coded in -1/1)}

\item{u}{Current value of u vector (ADMM optimization)}

\item{z}{Current value of z vector (ADMM optimization)}

\item{rho}{Tuning parameter for ADMM optimization}
}
\value{
Vector containing updated estimate of beta vector
}
\description{
The beta-update step requires optimizing a convex function. This version of the update function
uses R's built-it optim function to do this for us
}
