% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{b_updateC}
\alias{b_updateC}
\title{Update beta estimates using Newton-Raphson algorithm}
\usage{
b_updateC(X, y, u, z, rho, maxiter = 50L, toler = 1e-05, b = 0.5,
  alpha = 0.1)
}
\arguments{
\item{X}{Covariate matrix (no column for intercept)}

\item{y}{Vector of observations (coded in -1/1)}

\item{u}{Current value of u vector (ADMM optimization)}

\item{z}{Current value of z vector (ADMM optimization)}

\item{rho}{Tuning parameter for ADMM optimization}

\item{maxiter}{Maximum number of iterations}

\item{toler}{Convergence criterion}

\item{b}{Backtracking tuning parameter}

\item{alpha}{Backtracking tuning parameter}
}
\value{
Vector containing updated estimate of beta vector
}
\description{
The beta-update step requires optimizing a convex function. This version of the update function
uses a Newton-Raphson approach to minimizing the objective function.
}
